Add arithmetic workflows and restructure calculation plugins
Refactor the calculations package to remove the plugin directory. The
calculation job plugins that ship with `aiida-core` were being stored in
`aiida.calculations.plugins` but the `plugins` submodule is unnecessary
so it is removed. The old import is kept for now and will emit a
deprecation warning.

The `ArithemeticAddCalculation` is updated to simplify its usage as well
as to make it easier to understand the implementation, which is
important as it is used consistenly as an example in tutorials and the
documentation. The following changes are applied:

 * `metadata.options.resources` has a default
 * Unnecessary utility methods are removed to reduce amount of code
 * `calc_info.cmdline_params` are replaced with `calc_info.stdin_name`
 * Code will now work directly with bash instead of custom `add.sh` script

Finally, two workflows implementations are added for testing and
demonstration purposes. The `add_multiply` work function and the
`MultiplyAddWorkChain`. These will be used extensively in the how-to
guides of the documentation and the tutorials.
