+++ .ci/Dockerfile
@@ -92,9 +92,6 @@ COPY doubler.sh /usr/local/bin/
 # Use messed-up filename to test quoting robustness
 RUN mv /usr/local/bin/doubler.sh /usr/local/bin/d\"o\'ub\ ler.sh
 
-# Put the add script
-COPY add.sh /usr/local/bin/
-
 # add USER (no password); 1000 is the uid of the user in the jenkins docker
 RUN groupadd -g ${gid} jenkins && useradd -m -s /bin/bash -u ${uid} -g ${gid} jenkins
