+++ .github/config/add.yaml
@@ -4,6 +4,6 @@ description: add
 input_plugin: arithmetic.add
 on_computer: true
 computer: localhost
-remote_abs_path: PLACEHOLDER_REMOTE_ABS_PATH_ADD
+remote_abs_path: /bin/bash
 prepend_text: ' '
 append_text: ' '
