+++ .pylintrc
@@ -136,7 +136,7 @@ function-name-hint=(([a-z][a-z0-9_]{2,40})|(_[a-z0-9_]*))$
 function-rgx=(([a-z][a-z0-9_]{2,30})|(_[a-z0-9_]*))$
 
 # Good variable names which should always be accepted, separated by a comma
-good-names=i,j,k,ex,Run,_,pk,TemplatereplacerCalculation,ArithmeticAddCalculation
+good-names=x,y,z,i,j,k,ex,Run,_,pk,TemplatereplacerCalculation,ArithmeticAddCalculation
 
 # Include a hint for the correct naming format with invalid-name
 include-naming-hint=no
