+++ docs/source/concepts/calculations.rst
@@ -107,7 +107,7 @@ Let's imagine an external code that consists of a bash script that reads an inpu
 When run, this script reads the contents of a file called ``aiida.in`` and expects that it contains two integers.
 It will parse these into the variables ``x`` and ``y`` and then print their sum.
 When you want to run this 'code' through AiiDA, you need to tell *how* AiiDA should run it.
-The :py:class:`~aiida.calculations.plugins.arithmetic.add.ArithmeticAddCalculation` is a calculation job implementation that forms an interface to accomplish exactly that for the example bash script.
+The :py:class:`~aiida.calculations.arithmetic.add.ArithmeticAddCalculation` is a calculation job implementation that forms an interface to accomplish exactly that for the example bash script.
 A ``CalcJob`` implementation for a specific code, often referred to as a calculation plugin, essentially instructs the engine how it should be run.
 This includes how the necessary input files should be created based on the inputs that it receives, how the code executable should be called and what files should be retrieved when the calculation is complete.
 Note the files should be 'retrieved' because calculation jobs can be run not just on the localhost, but on any :ref:`computer that is configured in AiiDA<setup_computer>`, including remote machines accessible over for example SSH.
